refactor: move tests to the test dir
