+++ src/sys/signalfd.rs
@@ -155,34 +155,3 @@ impl Iterator for SignalFd {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn create_signalfd() {
-        let mask = SigSet::empty();
-        SignalFd::new(&mask).unwrap();
-    }
-
-    #[test]
-    fn create_signalfd_with_opts() {
-        let mask = SigSet::empty();
-        SignalFd::with_flags(
-            &mask,
-            SfdFlags::SFD_CLOEXEC | SfdFlags::SFD_NONBLOCK,
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn read_empty_signalfd() {
-        let mask = SigSet::empty();
-        let mut fd =
-            SignalFd::with_flags(&mask, SfdFlags::SFD_NONBLOCK).unwrap();
-
-        let res = fd.read_signal();
-        assert!(res.unwrap().is_none());
-    }
-}
