+++ src/sys/socket/sockopt.rs
@@ -1403,72 +1403,3 @@ impl<'a> Set<'a, OsString> for SetOsString<'a> {
     }
 }
 
-#[cfg(test)]
-mod test {
-    #[cfg(linux_android)]
-    #[test]
-    fn can_get_peercred_on_unix_socket() {
-        use super::super::*;
-
-        let (a, b) = socketpair(
-            AddressFamily::Unix,
-            SockType::Stream,
-            None,
-            SockFlag::empty(),
-        )
-        .unwrap();
-        let a_cred = getsockopt(&a, super::PeerCredentials).unwrap();
-        let b_cred = getsockopt(&b, super::PeerCredentials).unwrap();
-        assert_eq!(a_cred, b_cred);
-        assert_ne!(a_cred.pid(), 0);
-    }
-
-    #[test]
-    fn is_socket_type_unix() {
-        use super::super::*;
-
-        let (a, _b) = socketpair(
-            AddressFamily::Unix,
-            SockType::Stream,
-            None,
-            SockFlag::empty(),
-        )
-        .unwrap();
-        let a_type = getsockopt(&a, super::SockType).unwrap();
-        assert_eq!(a_type, SockType::Stream);
-    }
-
-    #[test]
-    fn is_socket_type_dgram() {
-        use super::super::*;
-
-        let s = socket(
-            AddressFamily::Inet,
-            SockType::Datagram,
-            SockFlag::empty(),
-            None,
-        )
-        .unwrap();
-        let s_type = getsockopt(&s, super::SockType).unwrap();
-        assert_eq!(s_type, SockType::Datagram);
-    }
-
-    #[cfg(any(target_os = "freebsd", target_os = "linux"))]
-    #[test]
-    fn can_get_listen_on_tcp_socket() {
-        use super::super::*;
-
-        let s = socket(
-            AddressFamily::Inet,
-            SockType::Stream,
-            SockFlag::empty(),
-            None,
-        )
-        .unwrap();
-        let s_listening = getsockopt(&s, super::AcceptConn).unwrap();
-        assert!(!s_listening);
-        listen(&s, 10).unwrap();
-        let s_listening2 = getsockopt(&s, super::AcceptConn).unwrap();
-        assert!(s_listening2);
-    }
-}
