+++ src/sys/statvfs.rs
@@ -148,20 +148,3 @@ pub fn fstatvfs<Fd: AsFd>(fd: Fd) -> Result<Statvfs> {
             .map(|_| Statvfs(stat.assume_init()))
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::sys::statvfs::*;
-    use std::fs::File;
-
-    #[test]
-    fn statvfs_call() {
-        statvfs(&b"/"[..]).unwrap();
-    }
-
-    #[test]
-    fn fstatvfs_call() {
-        let root = File::open("/").unwrap();
-        fstatvfs(&root).unwrap();
-    }
-}
