+++ src/sys/utsname.rs
@@ -62,24 +62,3 @@ fn cast_and_trim(slice: &[c_char]) -> &OsStr {
 
     OsStr::from_bytes(bytes)
 }
-
-#[cfg(test)]
-mod test {
-    #[cfg(target_os = "linux")]
-    #[test]
-    pub fn test_uname_linux() {
-        assert_eq!(super::uname().unwrap().sysname(), "Linux");
-    }
-
-    #[cfg(apple_targets)]
-    #[test]
-    pub fn test_uname_darwin() {
-        assert_eq!(super::uname().unwrap().sysname(), "Darwin");
-    }
-
-    #[cfg(target_os = "freebsd")]
-    #[test]
-    pub fn test_uname_freebsd() {
-        assert_eq!(super::uname().unwrap().sysname(), "FreeBSD");
-    }
-}
